Refactor the file structures.
